requirements.txt
attrs==22.2.0
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
clickclick==20.10.2
connexion==2.14.2
Flask==2.2.3
Flask-Authorize==0.2.6
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.0
pkgutil-resolve-name==1.3.10
psycopg2==2.8.4
pydantic==1.10.7
pyrsistent==0.19.3
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.7
SQLAlchemy-Utils==0.40.0
swagger-ui-bundle==0.0.9
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
